+++ Modules/Combat/CombatModule.lua
@@ -144,7 +144,7 @@ function CombatModule:ForfeitBattleHasStarted()
 
         local meetsForfeitThreshold = breedQuality >= (thresholdValues[forfeitThreshold] or 0)
 
-        if forfeitPromptUnless == _BattlePetCompletionist.Enums.ForfeitPromptUnless.NOT_MAX_COLLECTED or forfeitPromptUnless == _BattlePetCompletionist.Enums.ForfeitPromptUnless.NOT_MAX_RARE then
+        if forfeitPromptUnless == _BattlePetCompletionist.Enums.ForfeitPromptUnless.NOT_MAX_COLLECTED then
             if numCollected < limit and meetsForfeitThreshold then
                 upgradeFound = true
                 break
@@ -166,13 +166,20 @@ function CombatModule:ForfeitBattleHasStarted()
         end
 
         if forfeitPromptUnless == _BattlePetCompletionist.Enums.ForfeitPromptUnless.NOT_MAX_RARE then
-            -- Find the lowest quality of the pet that we own.
+            -- Not at max collected count is an upgrade.
+            if numCollected < limit and meetsForfeitThreshold then
+                upgradeFound = true
+                break
+            end
+
+            -- At max count, but the enemy pet can replace a lower-quality one.
             local lowestOwnedQuality = 4
             for _, myPetInfo in ipairs(myPets) do
                 if myPetInfo[2] < lowestOwnedQuality then
                     lowestOwnedQuality = myPetInfo[2]
                 end
             end
+
             if breedQuality > lowestOwnedQuality and meetsForfeitThreshold then
                 upgradeFound = true
                 break